+++ src/meta/stream/Stream.h
@@ -75,11 +75,11 @@ namespace MetaStream
             virtual int discNumber() const;
             virtual void setDiscNumber( int newDiscNumber );
 
-            virtual int length() const;
+            virtual qint64 length() const;
             virtual int filesize() const;
             virtual int sampleRate() const;
             virtual int bitrate() const;
-            
+
             virtual uint lastPlayed() const;
             virtual int playCount() const;
