+++ src/timecode/TimecodeMeta.h
@@ -47,7 +47,7 @@ public:
     BoundedPlaybackCapabilityImpl( TimecodeTrack * track )
         : m_track( track )
     {}
-    
+
     virtual qint64 startPosition();
     virtual qint64 endPosition();
 
@@ -123,7 +123,7 @@ public:
     virtual int rating() const;
     virtual void setRating ( int newRating );
 
-    virtual int length() const;
+    virtual qint64 length() const;
 
     virtual int filesize() const;
     virtual int sampleRate() const;
@@ -171,7 +171,7 @@ private:
     QString m_type;
     qint64 m_start;
     qint64 m_end;
-    int m_length;
+    qint64 m_length;
     int m_trackNumber;
     int m_discNumber;
     QString m_comment;
